Move autocomplete methods to index
